locator: split replication_strategy_type into its own header
db/config.hh included locator/abstract_replication_strategy.hh solely for
the replication_strategy_type enum class, but that transitively pulled in
tablets.hh, raft/raft.hh, token_metadata.hh, and topology.hh (~3600 lines)
into every one of ~136 direct consumers of db/config.hh.

Fix: add locator/replication_strategy_type.hh containing only the enum and
its fmt::formatter declaration. Switch db/config.hh to include that instead.
Update the ~8 TUs that were relying on the transitive include to include
abstract_replication_strategy.hh directly.

(cherry picked from commit 48b2fd0af73b5179acd74e4afd875800c4601914,
originally part of scylladb/scylladb#29009)
