+++ dht/range_streamer.cc
@@ -14,6 +14,7 @@
 #include "utils/log.hh"
 #include "streaming/stream_plan.hh"
 #include "db/config.hh"
+#include "locator/abstract_replication_strategy.hh"
 #include <fmt/ranges.h>
 #include <seastar/core/semaphore.hh>
 #include <seastar/core/sleep.hh>
