+++ repair/row_level.cc
@@ -47,6 +47,8 @@
 #include <seastar/core/coroutine.hh>
 #include <seastar/coroutine/all.hh>
 #include <seastar/coroutine/as_future.hh>
+#include "db/config.hh"
+#include "locator/abstract_replication_strategy.hh"
 #include "db/system_keyspace.hh"
 #include "service/storage_proxy.hh"
 #include "db/batchlog_manager.hh"
