+++ service/storage_service.cc
@@ -89,6 +89,7 @@
 #include "sstables/sstables.hh"
 #include "sstables/sstables_manager.hh"
 #include "db/config.hh"
+#include "locator/abstract_replication_strategy.hh"
 #include "db/schema_tables.hh"
 #include "db/view/view_builder.hh"
 #include "replica/database.hh"
